requirements.txt
discord.py
python-dotenv
Flask
psycopg2-binary
